# specs/1-todo-backend-fastapi/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Backend API
  version: 1.0.0
  description: API for managing user tasks with JWT authentication
servers:
  - url: http://localhost:8000/api/{user_id}
    description: Development server
paths:
  /tasks:
    get:
      summary: Get all tasks for a user
      description: Retrieve all tasks belonging to the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier from JWT token
          schema:
            type: string
      responses:
        '200':
          description: Successful retrieval of user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Attempt to access another user's tasks
    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier from JWT token
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTask'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Attempt to access another user's tasks
  /tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier from JWT token
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      responses:
        '200':
          description: Successful retrieval of task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Task belongs to another user
        '404':
          description: Task not found
    put:
      summary: Update a task
      description: Update all details of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier from JWT token
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatedTask'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Task belongs to another user
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      description: Remove a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier from JWT token
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Task belongs to another user
        '404':
          description: Task not found
  /tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggle the completion status of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier from JWT token
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: string
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Task belongs to another user
        '404':
          description: Task not found
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        user_id:
          type: string
          description: Identifier of the user who owns this task
        title:
          type: string
          description: Title of the task
          minLength: 1
        description:
          type: string
          description: Optional description of the task
        completed:
          type: boolean
          description: Whether the task is completed
    NewTask:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the task
          minLength: 1
        description:
          type: string
          description: Optional description of the task
        completed:
          type: boolean
          description: Whether the task is completed (defaults to false)
    UpdatedTask:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the task
          minLength: 1
        description:
          type: string
          description: Updated description of the task
        completed:
          type: boolean
          description: Updated completion status
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []